requirements.txt
Flask==2.0.2
Flask-SocketIO==5.1.1
requests==2.26.0
python-dotenv==0.19.2
eventlet==0.33.0
